// stories/types.ts
export type Fruit = {
  id: number;
  name: string;
};

export type Color = {
  id: number;
  name: string;
  color: string;
};

export type Song = {
  id: number;
  artist: string;
  song: string;
  year: number;
};

export type Sport = {
  guid: string;
  name: string;
};

export type User = {
  name: string;
  age: number;
  email: string;
};

export type Language = {
  name: string;
  type: string;
  year: number;
};

export type Currency = {
  code: string;
  label: string;
  symbol: string;
};